requirements.txt
django-model-utils
django-bootstrap3
fabric
django-celery
channels
redis
